1.0.0/requirements.txt
# For PyTorch, install separately: pip install torch torchvision torchaudio
transformers>=4.35.0
openai-whisper>=20231117
librosa>=0.10.0
soundfile>=0.12.0
scikit-learn>=1.3.0 
pydub>=0.25.0
pyaudio>=0.2.11
numpy>=1.24.0
scipy>=1.10.0
noisereduce>=3.0.3
PySide6>=6.4.0
pyobjc>=10.2
cryptography>=41.0.0
pathlib2>=2.3.0
python-magic>=0.4.27 
psutil>=7.1.1
certifi>=2023.6.15
python-docx>=0.8.11
openpyxl>=3.1.0
langdetect>=1.0.9
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
py2app>=0.28.8
setuptools>=68.0.0